requirements.txt
Flask==1.0.2
opencv-python-headless==4.5.5.64
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Werkzeug==0.15.5
tensorflow-cpu==2.4.1
Keras==2.4.0
Pillow>=5.4.0
numpy==1.19.2
scikit-learn==0.22.2
pandas==0.25.1
h5py==2.10.0
pyyaml==6.0
packaging==21.0
matplotlib==3.3.4

